+++ nope/lib/models/User.php
@@ -84,17 +84,22 @@ class User extends \Nope\Model {
     $this->model->password = $hashedPassword;
   }
 
-  static function authenticate($username, $password) {
+  static function login($username, $password) {
     $user = self::findByUsername($username);
     if($user && v::identical($user->password)->validate(\Nope\Utils::hashPassword($password,$user->salt)) && $user->enabled==1) {
-      $user->lastLoginDate = new \DateTime();
-      $user->resetCode = null;
-      $user->save();
-      $user->saveInSession();
+      self::authenticate($user);
       return true;
     } else {
       return false;
     }
+  }
+
+  static function authenticate($user) {
+    $user->lastLoginDate = new \DateTime();
+    $user->resetCode = null;
+    $user->save();
+    $user->saveInSession();
+    return true;
   }
 
   static function getAuthenticated() {